Move to common compo, and more stories
