+++ src/components/Tooltips/TooltipsWindow/TooltipsWindow.tsx
@@ -1,4 +1,3 @@
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 const Cont = styled.div`
