+++ src/components/common/CheckBoxColor/CheckBoxColor.tsx
