+++ src/components/common/CheckBoxSwitch/CheckBoxSwitch.tsx
