+++ src/stories/CheckBox.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import CheckBox from '../components/CheckBox/CheckBox';
+import CheckBox from '../components/common/CheckBox/CheckBox';
 
 const meta: Meta<typeof CheckBox> = {
 	component: CheckBox,
@@ -10,5 +10,14 @@ export default meta;
 type Story = StoryObj<typeof CheckBox>;
 
 export const Default: Story = {
-	args: {},
+	args: {
+		id: 'checkbox',
+		name: 'checkbox',
+	},
+};
+
+export const State: Story = {
+	args: {
+		checked: true,
+	},
 };
