+++ src/stories/CheckBoxSwitch.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import CheckBoxSwitch from '../components/CheckBoxSwitch/CheckBoxSwitch';
+import CheckBoxSwitch from '../components/common/CheckBoxSwitch/CheckBoxSwitch';
 
 const meta: Meta<typeof CheckBoxSwitch> = {
 	component: CheckBoxSwitch,
@@ -10,5 +10,14 @@ export default meta;
 type Story = StoryObj<typeof CheckBoxSwitch>;
 
 export const Default: Story = {
-	args: {},
+	args: {
+		id: 'checkbox',
+		name: 'checkbox',
+	},
+};
+
+export const State: Story = {
+	args: {
+		checked: true,
+	},
 };
